refactor(allowed recipients): move some tests into separate files
