+++ tests/test_allowed_recipients_motions.py
@@ -1,0 +1,353 @@
+from brownie.network import chain
+from brownie import reverts
+from typing import List
+
+from eth_abi import encode_single
+from utils.evm_script import encode_call_script, encode_calldata
+
+from utils.config import get_network_name
+
+from utils.lido import create_voting, execute_voting
+
+from utils.test_helpers import (
+    assert_event_exists,
+    access_revert_message,
+    SET_LIMIT_PARAMETERS_ROLE,
+    UPDATE_SPENT_AMOUNT_ROLE,
+    ADD_RECIPIENT_TO_ALLOWED_LIST_ROLE,
+    REMOVE_RECIPIENT_FROM_ALLOWED_LIST_ROLE,
+)
+
+import constants
+
+
+MAX_SECONDS_IN_MONTH = 31 * 24 * 60 * 60
+
+
+def add_recipient(
+    recipient, recipient_title, easy_track, add_recipient_factory, allowed_recipients_registry
+):
+    call_data = encode_calldata("(address,string)", [recipient.address, recipient_title])
+    assert add_recipient_factory.decodeEVMScriptCallData(call_data) == [
+        recipient.address,
+        recipient_title,
+    ]
+
+    tx = easy_track.createMotion(
+        add_recipient_factory,
+        call_data,
+        {"from": add_recipient_factory.trustedCaller()},
+    )
+
+    chain.sleep(constants.MIN_MOTION_DURATION + 100)
+
+    tx = easy_track.enactMotion(
+        easy_track.getMotions()[0][0],
+        tx.events["MotionCreated"]["_evmScriptCallData"],
+        {"from": recipient},
+    )
+    assert_event_exists(
+        tx,
+        "RecipientAdded",
+        {"_recipient": recipient, "_title": recipient_title},
+    )
+    assert allowed_recipients_registry.isRecipientAllowed(recipient)
+
+
+def remove_recipient(recipient, easy_track, remove_recipient_factory, allowed_recipients_registry):
+    call_data = encode_single("(address)", [recipient.address])
+    assert remove_recipient_factory.decodeEVMScriptCallData(call_data) == recipient.address
+
+    tx = easy_track.createMotion(
+        remove_recipient_factory,
+        call_data,
+        {"from": remove_recipient_factory.trustedCaller()},
+    )
+
+    assert tx.events["MotionCreated"]["_evmScriptCallData"] == "0x" + call_data.hex()
+
+    chain.sleep(constants.MIN_MOTION_DURATION + 100)
+
+    tx = easy_track.enactMotion(
+        easy_track.getMotions()[0][0],
+        call_data,
+        {"from": recipient},
+    )
+    assert_event_exists(
+        tx,
+        "RecipientRemoved",
+        {"_recipient": recipient},
+    )
+    assert not allowed_recipients_registry.isRecipientAllowed(recipient)
+
+
+def set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent):
+    """Do Aragon voting to set limit parameters to the allowed recipients registry"""
+    set_limit_parameters_voting_id, _ = create_voting(
+        evm_script=encode_call_script(
+            [
+                (
+                    allowed_recipients_registry.address,
+                    allowed_recipients_registry.setLimitParameters.encode_input(
+                        period_limit,
+                        period_duration,
+                    ),
+                ),
+            ]
+        ),
+        description="Set limit parameters",
+        network=get_network_name(),
+        tx_params={"from": agent},
+    )
+
+    # execute voting to add permissions to EVM script executor to create payments
+    execute_voting(set_limit_parameters_voting_id, get_network_name())
+
+    allowed_recipients_registry.getLimitParameters() == (period_limit, period_duration)
+
+
+def create_top_up_motion(recipients: List[str], amounts: List[int], easy_track, top_up_factory):
+    script_call_data = encode_single(
+        "(address[],uint256[])",
+        [recipients, amounts],
+    )
+    tx = easy_track.createMotion(
+        top_up_factory,
+        script_call_data,
+        {"from": top_up_factory.trustedCaller()},
+    )
+    motion_id = tx.events["MotionCreated"]["_motionId"]
+    return motion_id, script_call_data
+
+
+def do_payout_to_allowed_recipients(recipients, amounts, easy_track, top_up_factory):
+    motion_id, script_call_data = create_top_up_motion(
+        recipients, amounts, easy_track, top_up_factory
+    )
+
+    chain.sleep(constants.MIN_MOTION_DURATION + 100)
+
+    easy_track.enactMotion(
+        motion_id,
+        script_call_data,
+        {"from": recipients[0]},
+    )
+
+
+def test_add_same_recipient_twice(entire_allowed_recipients_setup, accounts):
+    (
+        easy_track,
+        _,  # evm_script_executor,
+        allowed_recipients_registry,
+        _,  # top_up_factory,
+        add_recipient_factory,
+        _,  # remove_recipient_factory,
+    ) = entire_allowed_recipients_setup
+
+    recipient = accounts[8]
+    recipient_title = "New Allowed Recipient"
+
+    add_recipient(
+        recipient, recipient_title, easy_track, add_recipient_factory, allowed_recipients_registry
+    )
+
+    with reverts("ALLOWED_RECIPIENT_ALREADY_ADDED"):
+        add_recipient(
+            recipient,
+            recipient_title,
+            easy_track,
+            add_recipient_factory,
+            allowed_recipients_registry,
+        )
+
+
+def test_remove_not_added_recipient(entire_allowed_recipients_setup, accounts):
+    (
+        easy_track,
+        _,  # evm_script_executor,
+        allowed_recipients_registry,
+        _,  # top_up_factory,
+        _,  # add_recipient_factory,
+        remove_recipient_factory,
+    ) = entire_allowed_recipients_setup
+
+    recipient = accounts[8]
+    assert not allowed_recipients_registry.isRecipientAllowed(recipient)
+
+    with reverts("ALLOWED_RECIPIENT_NOT_FOUND"):
+        remove_recipient(
+            recipient, easy_track, remove_recipient_factory, allowed_recipients_registry
+        )
+
+
+def test_add_remove_recipient_happy_path(entire_allowed_recipients_setup, accounts, stranger):
+    (
+        easy_track,
+        _,  # evm_script_executor,
+        allowed_recipients_registry,
+        _,  # top_up_factory,
+        add_recipient_factory,
+        remove_recipient_factory,
+    ) = entire_allowed_recipients_setup
+
+    recipient = accounts[8]
+    recipient_title = "New Allowed Recipient"
+
+    add_recipient(
+        recipient, recipient_title, easy_track, add_recipient_factory, allowed_recipients_registry
+    )
+    assert len(easy_track.getMotions()) == 0
+    assert allowed_recipients_registry.getAllowedRecipients() == [recipient]
+    assert not allowed_recipients_registry.isRecipientAllowed(stranger)
+
+    remove_recipient(recipient, easy_track, remove_recipient_factory, allowed_recipients_registry)
+    assert len(allowed_recipients_registry.getAllowedRecipients()) == 0
+
+
+def test_second_top_up_motion_exceeds_limit_in_same_period(
+    entire_allowed_recipients_setup_with_two_recipients,
+    agent,
+):
+    """Revert 2nd payout which together with 1st payout exceed limits in the same period"""
+    (
+        easy_track,
+        _,  # evm_script_executor,
+        allowed_recipients_registry,
+        top_up_factory,
+        _,  # add_recipient_factory,
+        _,  # remove_recipient_factory,
+        recipient1,
+        recipient2,
+    ) = entire_allowed_recipients_setup_with_two_recipients
+
+    period_limit, period_duration = 100 * 10**18, 6
+    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
+
+    recipients = list(map(lambda x: x.address, [recipient1, recipient2]))
+    do_payout_to_allowed_recipients(recipients, [int(3e18), int(90e18)], easy_track, top_up_factory)
+
+    with reverts("SUM_EXCEEDS_SPENDABLE_BALANCE"):
+        do_payout_to_allowed_recipients(
+            recipients, [int(5e18), int(4e18)], easy_track, top_up_factory
+        )
+
+
+def test_limit_is_renewed_in_next_period(
+    entire_allowed_recipients_setup_with_two_recipients, agent
+):
+    """Check limit is renewed in the next period"""
+    (
+        easy_track,
+        _,  # evm_script_executor,
+        allowed_recipients_registry,
+        top_up_factory,
+        _,  # add_recipient_factory,
+        _,  # remove_recipient_factory,
+        recipient1,
+        recipient2,
+    ) = entire_allowed_recipients_setup_with_two_recipients
+
+    period_limit, period_duration = 100 * 10**18, 1
+    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
+
+    recipients = list(map(lambda x: x.address, [recipient1, recipient2]))
+    do_payout_to_allowed_recipients(recipients, [int(3e18), int(90e18)], easy_track, top_up_factory)
+
+    chain.sleep(period_duration * MAX_SECONDS_IN_MONTH)
+
+    second_payout = [int(5e18), int(4e18)]
+    do_payout_to_allowed_recipients(recipients, second_payout, easy_track, top_up_factory)
+    assert allowed_recipients_registry.getPeriodState()[0] == sum(second_payout)
+
+
+def test_both_motions_enacted_next_period_second_exceeds_limit(
+    entire_allowed_recipients_setup_with_two_recipients, agent
+):
+    """
+    Two motion created this period, the first enacted next period, the second
+    is reverted due to the limit in the second period
+    """
+    (
+        easy_track,
+        _,  # evm_script_executor,
+        allowed_recipients_registry,
+        top_up_factory,
+        _,  # add_recipient_factory,
+        _,  # remove_recipient_factory,
+        recipient1,
+        recipient2,
+    ) = entire_allowed_recipients_setup_with_two_recipients
+    recipients = [recipient1.address, recipient2.address]
+
+    period_limit, period_duration = 100 * 10**18, 1
+    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
+
+    payout1 = [int(40e18), int(30e18)]
+    payout2 = [int(30e18), int(20e18)]
+    motion1_id, motion1_calldata = create_top_up_motion(
+        recipients, payout1, easy_track, top_up_factory
+    )
+    motion2_id, motion2_calldata = create_top_up_motion(
+        recipients, payout2, easy_track, top_up_factory
+    )
+
+    chain.sleep(period_duration * MAX_SECONDS_IN_MONTH)
+
+    easy_track.enactMotion(motion1_id, motion1_calldata, {"from": recipient1})
+
+    with reverts("SUM_EXCEEDS_SPENDABLE_BALANCE"):
+        easy_track.enactMotion(motion2_id, motion2_calldata, {"from": recipient1})
+
+
+def test_create_top_up_motion_above_limits(
+    entire_allowed_recipients_setup_with_two_recipients, agent
+):
+    (
+        easy_track,
+        _,  # evm_script_executor,
+        allowed_recipients_registry,
+        top_up_factory,
+        _,  # add_recipient_factory,
+        _,  # remove_recipient_factory,
+        recipient1,
+        _,  # recipient2,
+    ) = entire_allowed_recipients_setup_with_two_recipients
+
+    period_limit, period_duration = 100 * 10**18, 1
+    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
+
+    with reverts("SUM_EXCEEDS_SPENDABLE_BALANCE"):
+        create_top_up_motion([recipient1.address], [period_limit + 1], easy_track, top_up_factory)
+
+
+def test_limit_decreased_while_motion_is_in_flight(
+    entire_allowed_recipients_setup_with_two_recipients, agent
+):
+    (
+        easy_track,
+        _,  # evm_script_executor,
+        allowed_recipients_registry,
+        top_up_factory,
+        _,  # add_recipient_factory,
+        _,  # remove_recipient_factory,
+        recipient1,
+        _,  # recipient2,
+    ) = entire_allowed_recipients_setup_with_two_recipients
+
+    period_limit, period_duration = 100 * 10**18, 1
+    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
+
+    motion_id, motion_calldata = create_top_up_motion(
+        [recipient1.address], [period_limit], easy_track, top_up_factory
+    )
+
+    set_limit_parameters(period_limit // 2, period_duration, allowed_recipients_registry, agent)
+
+    chain.sleep(constants.MIN_MOTION_DURATION + 100)
+
+    with reverts("SUM_EXCEEDS_SPENDABLE_BALANCE"):
+        easy_track.enactMotion(
+            motion_id,
+            motion_calldata,
+            {"from": recipient1},
+        )
