+++ tests/test_allowed_recipients_registry.py
@@ -237,171 +237,6 @@ def test_allowed_recipients_registry_roles(
             registry.updateSpentAmount(1, {"from": caller})
 
 
-def test_add_same_recipient_twice(entire_allowed_recipients_setup, accounts):
-    (
-        easy_track,
-        _,  # evm_script_executor,
-        allowed_recipients_registry,
-        _,  # top_up_factory,
-        add_recipient_factory,
-        _,  # remove_recipient_factory,
-    ) = entire_allowed_recipients_setup
-
-    recipient = accounts[8]
-    recipient_title = "New Allowed Recipient"
-
-    add_recipient(
-        recipient, recipient_title, easy_track, add_recipient_factory, allowed_recipients_registry
-    )
-
-    with reverts("ALLOWED_RECIPIENT_ALREADY_ADDED"):
-        add_recipient(
-            recipient,
-            recipient_title,
-            easy_track,
-            add_recipient_factory,
-            allowed_recipients_registry,
-        )
-
-
-def test_remove_not_added_recipient(entire_allowed_recipients_setup, accounts):
-    (
-        easy_track,
-        _,  # evm_script_executor,
-        allowed_recipients_registry,
-        _,  # top_up_factory,
-        _,  # add_recipient_factory,
-        remove_recipient_factory,
-    ) = entire_allowed_recipients_setup
-
-    recipient = accounts[8]
-    assert not allowed_recipients_registry.isRecipientAllowed(recipient)
-
-    with reverts("ALLOWED_RECIPIENT_NOT_FOUND"):
-        remove_recipient(
-            recipient, easy_track, remove_recipient_factory, allowed_recipients_registry
-        )
-
-
-def test_add_remove_recipient_happy_path(entire_allowed_recipients_setup, accounts, stranger):
-    (
-        easy_track,
-        _,  # evm_script_executor,
-        allowed_recipients_registry,
-        _,  # top_up_factory,
-        add_recipient_factory,
-        remove_recipient_factory,
-    ) = entire_allowed_recipients_setup
-
-    recipient = accounts[8]
-    recipient_title = "New Allowed Recipient"
-
-    add_recipient(
-        recipient, recipient_title, easy_track, add_recipient_factory, allowed_recipients_registry
-    )
-    assert len(easy_track.getMotions()) == 0
-    assert allowed_recipients_registry.getAllowedRecipients() == [recipient]
-    assert not allowed_recipients_registry.isRecipientAllowed(stranger)
-
-    remove_recipient(recipient, easy_track, remove_recipient_factory, allowed_recipients_registry)
-    assert len(allowed_recipients_registry.getAllowedRecipients()) == 0
-
-
-def test_second_top_up_motion_exceeds_limit_in_same_period(
-    entire_allowed_recipients_setup_with_two_recipients,
-    agent,
-):
-    """Revert 2nd payout which together with 1st payout exceed limits in the same period"""
-    (
-        easy_track,
-        _,  # evm_script_executor,
-        allowed_recipients_registry,
-        top_up_factory,
-        _,  # add_recipient_factory,
-        _,  # remove_recipient_factory,
-        recipient1,
-        recipient2,
-    ) = entire_allowed_recipients_setup_with_two_recipients
-
-    period_limit, period_duration = 100 * 10**18, 6
-    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
-
-    recipients = list(map(lambda x: x.address, [recipient1, recipient2]))
-    do_payout_to_allowed_recipients(recipients, [int(3e18), int(90e18)], easy_track, top_up_factory)
-
-    with reverts("SUM_EXCEEDS_SPENDABLE_BALANCE"):
-        do_payout_to_allowed_recipients(
-            recipients, [int(5e18), int(4e18)], easy_track, top_up_factory
-        )
-
-
-def test_limit_is_renewed_in_next_period(
-    entire_allowed_recipients_setup_with_two_recipients, agent
-):
-    """Check limit is renewed in the next period"""
-    (
-        easy_track,
-        _,  # evm_script_executor,
-        allowed_recipients_registry,
-        top_up_factory,
-        _,  # add_recipient_factory,
-        _,  # remove_recipient_factory,
-        recipient1,
-        recipient2,
-    ) = entire_allowed_recipients_setup_with_two_recipients
-
-    period_limit, period_duration = 100 * 10**18, 1
-    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
-
-    recipients = list(map(lambda x: x.address, [recipient1, recipient2]))
-    do_payout_to_allowed_recipients(recipients, [int(3e18), int(90e18)], easy_track, top_up_factory)
-
-    chain.sleep(period_duration * MAX_SECONDS_IN_MONTH)
-
-    second_payout = [int(5e18), int(4e18)]
-    do_payout_to_allowed_recipients(recipients, second_payout, easy_track, top_up_factory)
-    assert allowed_recipients_registry.getPeriodState()[0] == sum(second_payout)
-
-
-def test_both_motions_enacted_next_period_second_exceeds_limit(
-    entire_allowed_recipients_setup_with_two_recipients, agent
-):
-    """
-    Two motion created this period, the first enacted next period, the second
-    is reverted due to the limit in the second period
-    """
-    (
-        easy_track,
-        _,  # evm_script_executor,
-        allowed_recipients_registry,
-        top_up_factory,
-        _,  # add_recipient_factory,
-        _,  # remove_recipient_factory,
-        recipient1,
-        recipient2,
-    ) = entire_allowed_recipients_setup_with_two_recipients
-    recipients = [recipient1.address, recipient2.address]
-
-    period_limit, period_duration = 100 * 10**18, 1
-    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
-
-    payout1 = [int(40e18), int(30e18)]
-    payout2 = [int(30e18), int(20e18)]
-    motion1_id, motion1_calldata = create_top_up_motion(
-        recipients, payout1, easy_track, top_up_factory
-    )
-    motion2_id, motion2_calldata = create_top_up_motion(
-        recipients, payout2, easy_track, top_up_factory
-    )
-
-    chain.sleep(period_duration * MAX_SECONDS_IN_MONTH)
-
-    easy_track.enactMotion(motion1_id, motion1_calldata, {"from": recipient1})
-
-    with reverts("SUM_EXCEEDS_SPENDABLE_BALANCE"):
-        easy_track.enactMotion(motion2_id, motion2_calldata, {"from": recipient1})
-
-
 def test_limits_checker_access_restriction(
     owner, lego_program, stranger, LimitsCheckerWrapper, easy_track, bokkyPooBahsDateTimeContract
 ):
@@ -630,119 +465,3 @@ def test_limits_checker_general(
 
     with reverts("SUM_EXCEEDS_SPENDABLE_BALANCE"):
         limits_checker.updateSpentAmount(1, {"from": script_executor})
-
-
-def test_top_up_factory_evm_script_validation(
-    AllowedRecipientsRegistry,
-    TopUpAllowedRecipients,
-    owner,
-    finance,
-    ldo,
-    easy_track,
-    voting,
-    bokkyPooBahsDateTimeContract,
-    stranger,
-):
-    deployer = owner
-    trusted_caller = owner
-    manager = accounts[7]
-    recipient = deployer.address
-
-    registry = deployer.deploy(
-        AllowedRecipientsRegistry,
-        voting,
-        [manager],
-        [manager],
-        [manager],
-        [manager],
-        bokkyPooBahsDateTimeContract,
-    )
-    registry.addRecipient(recipient, "Test Recipient", {"from": manager})
-    registry.setLimitParameters(int(100e18), 12, {"from": manager})
-
-    top_up_factory = deployer.deploy(
-        TopUpAllowedRecipients, trusted_caller, registry, finance, ldo, easy_track
-    )
-
-    def make_call_data(recipients, amounts):
-        return encode_single("(address[],uint256[])", [recipients, amounts])
-
-    with reverts("EMPTY_DATA"):
-        top_up_factory.createEVMScript(trusted_caller, make_call_data([], []))
-
-    with reverts("LENGTH_MISMATCH"):
-        top_up_factory.createEVMScript(trusted_caller, make_call_data([recipient], []))
-
-    with reverts("LENGTH_MISMATCH"):
-        top_up_factory.createEVMScript(trusted_caller, make_call_data([], [123]))
-
-    with reverts("ZERO_AMOUNT"):
-        top_up_factory.createEVMScript(trusted_caller, make_call_data([recipient], [0]))
-
-    with reverts("ZERO_AMOUNT"):
-        top_up_factory.createEVMScript(
-            trusted_caller, make_call_data([recipient, recipient], [123, 0])
-        )
-
-    with reverts("RECIPIENT_NOT_ALLOWED"):
-        top_up_factory.createEVMScript(trusted_caller, make_call_data([stranger.address], [123]))
-
-    payout = int(1e18)
-    call_data = make_call_data([recipient], [payout])
-    evm_script = top_up_factory.createEVMScript(trusted_caller, call_data)
-    assert top_up_factory.decodeEVMScriptCallData(call_data) == ([recipient], [payout])
-    assert "Top up allowed recipients".encode("utf-8").hex() in str(evm_script)
-
-
-def test_create_top_up_motion_above_limits(
-    entire_allowed_recipients_setup_with_two_recipients, agent
-):
-    (
-        easy_track,
-        _,  # evm_script_executor,
-        allowed_recipients_registry,
-        top_up_factory,
-        _,  # add_recipient_factory,
-        _,  # remove_recipient_factory,
-        recipient1,
-        _,  # recipient2,
-    ) = entire_allowed_recipients_setup_with_two_recipients
-
-    period_limit, period_duration = 100 * 10**18, 1
-    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
-
-    with reverts("SUM_EXCEEDS_SPENDABLE_BALANCE"):
-        create_top_up_motion([recipient1.address], [period_limit + 1], easy_track, top_up_factory)
-
-
-def test_limit_decreased_while_motion_is_in_flight(
-    entire_allowed_recipients_setup_with_two_recipients, agent
-):
-    (
-        easy_track,
-        _,  # evm_script_executor,
-        allowed_recipients_registry,
-        top_up_factory,
-        _,  # add_recipient_factory,
-        _,  # remove_recipient_factory,
-        recipient1,
-        _,  # recipient2,
-    ) = entire_allowed_recipients_setup_with_two_recipients
-
-    period_limit, period_duration = 100 * 10**18, 1
-    set_limit_parameters(period_limit, period_duration, allowed_recipients_registry, agent)
-
-    motion_id, motion_calldata = create_top_up_motion(
-        [recipient1.address], [period_limit], easy_track, top_up_factory
-    )
-
-    set_limit_parameters(period_limit // 2, period_duration, allowed_recipients_registry, agent)
-
-    chain.sleep(constants.MIN_MOTION_DURATION + 100)
-
-    with reverts("SUM_EXCEEDS_SPENDABLE_BALANCE"):
-        easy_track.enactMotion(
-            motion_id,
-            motion_calldata,
-            {"from": recipient1},
-        )
